Update vgchartzfull.py
This version allows for a restart it will determine the correct page to start at. It is also gathering the list of games in a series and the list of platforms if platform is listed as All, they are divided by |. There is a sleep timer to prevent 429 errors. Also it is dumbing the variables after each page to keep memory free and also prevent crashes.